pyproject.toml
[project]
name = "feishu-weekend-buddy-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.0.7",
    "langchain-deepseek>=1.0.1",
    "langchain-openai==1.1.6",
    "langgraph==1.0.5",
    "langgraph-prebuilt==1.0.5",
    "lark-oapi>=1.5.3",
    "openai==2.14.0",
    "python-dotenv>=1.2.2",
    "python-socks>=2.8.1",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]
